+++ tests/test_checker/test_noqa.py
@@ -181,6 +181,7 @@ SHOULD_BE_RAISED = types.MappingProxyType({
     'WPS608': 1,
     'WPS609': 1,
     'WPS610': 1,
+    'WPS611': 1,
 })