+++ wemake_python_styleguide/visitors/ast/functions.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 
 import ast
-from itertools import zip_longest
 from typing import ClassVar, Dict, List, Optional, Union
 
 from typing_extensions import final
@@ -18,6 +17,7 @@ from wemake_python_styleguide.logic import (
     operators,
     walk,
 )
+from wemake_python_styleguide.logic.arguments import function_args
 from wemake_python_styleguide.logic.naming import access
 from wemake_python_styleguide.types import AnyFunctionDef, AnyNodes
 from wemake_python_styleguide.violations.best_practices import (
@@ -236,68 +236,6 @@ class UselessLambdaDefinitionVisitor(base.BaseNodeVisitor):
         self._check_useless_lambda(node)
         self.generic_visit(node)
 
-    def _have_same_kwarg(self, node: ast.Lambda, call: ast.Call) -> bool:
-        kwarg_name: Optional[str] = None
-        for keyword in call.keywords:
-            # `a=1` vs `**kwargs`:
-            # {'arg': 'a', 'value': <_ast.Num object at 0x1027882b0>}
-            # {'arg': None, 'value': <_ast.Name object at 0x102788320>}
-            if keyword.arg is None:
-                if isinstance(keyword.value, ast.Name):
-                    kwarg_name = keyword.value.id
-                else:  # We can judge on things like `**{}`
-                    return False
-        if node.args.kwarg and kwarg_name:
-            return node.args.kwarg.arg == kwarg_name
-        return node.args.kwarg == kwarg_name
-
-    def _have_same_vararg(self, node: ast.Lambda, call: ast.Call) -> bool:
-        vararg_name: Optional[str] = None
-        for ar in call.args:
-            # 'args': [<_ast.Starred object at 0x10d77a3c8>]
-            if isinstance(ar, ast.Starred):
-                if isinstance(ar.value, ast.Name):
-                    vararg_name = ar.value.id
-                else:  # We can judge on things like `*[]`
-                    return False
-        if vararg_name and node.args.vararg:
-            return node.args.vararg.arg == vararg_name
-        return node.args.vararg == vararg_name
-
-    def _have_same_args(self, node: ast.Lambda, call: ast.Call) -> bool:
-        paired_arguments = zip_longest(call.args, node.args.args)
-        for call_arg, lambda_arg in paired_arguments:
-            if isinstance(call_arg, ast.Starred):
-                if isinstance(lambda_arg, ast.arg):
-                    return False
-            elif isinstance(call_arg, ast.Name):
-                if not lambda_arg or call_arg.id != lambda_arg.arg:
-                    return False
-            else:
-                return False
-        return True
-
-    def _have_same_kw_args(self, node: ast.Lambda, call: ast.Call) -> bool:
-        prepared_kw_args = {
-            kw.arg: kw
-            for kw in call.keywords
-            if isinstance(kw.value, ast.Name) and kw.arg == kw.value.id
-        }
-
-        real_kw_args = [
-            # We need to remove ** args from here:
-            kw for kw in call.keywords
-            if not (isinstance(kw.value, ast.Name) and kw.arg is None)
-        ]
-
-        for lambda_arg in node.args.kwonlyargs:
-            lambda_arg_name = getattr(lambda_arg, 'arg', None)
-            call_arg = prepared_kw_args.get(lambda_arg_name)
-
-            if lambda_arg and not call_arg:
-                return False
-        return len(real_kw_args) == len(node.args.kwonlyargs)
-
     def _check_useless_lambda(self, node: ast.Lambda) -> None:
         if not isinstance(node.body, ast.Call):
             return
@@ -312,14 +250,7 @@ class UselessLambdaDefinitionVisitor(base.BaseNodeVisitor):
             # `kw_defaults` can have [None, ...] items.
             return
 
-        same_vararg = self._have_same_vararg(node, node.body)
-        same_kwarg = self._have_same_kwarg(node, node.body)
-        if not same_vararg or not same_kwarg:
-            return
-
-        same_args = self._have_same_args(node, node.body)
-        same_kw_args = self._have_same_kw_args(node, node.body)
-        if not same_args or not same_kw_args:
+        if not function_args.is_call_matched_by_arguments(node, node.body):
             return
 
         self.add_violation(UselessLambdaViolation(node))